+++ decidim-verifications/spec/queries/decidim/verifications/authorizations_spec.rb
@@ -4,18 +4,21 @@ require "spec_helper"
 
 describe Decidim::Verifications::Authorizations do
   let(:name) { "some_method" }
-  let(:user) { create(:user) }
+  let(:user) { create(:user, organization: organization) }
+  let(:user2) { create(:user, organization: organization) }
+  let(:user3) { create(:user, organization: organization) }
+  let(:organization) { create :organization }
 
   let!(:granted_for_user_and_method) do
     create(:authorization, :granted, name: name, user: user)
   end
 
   let!(:pending_for_other_user) do
-    create(:authorization, :pending, name: name)
+    create(:authorization, :pending, name: name, user: user2)
   end
 
   let!(:granted_for_other_user) do
-    create(:authorization, :granted, name: name)
+    create(:authorization, :granted, name: name, user: user3)
   end
 
   let!(:pending_for_other_method) do
@@ -27,23 +30,28 @@ describe Decidim::Verifications::Authorizations do
   end
 
   let!(:pending_for_other_user_and_method) do
-    create(:authorization, :pending)
+    create(:authorization, :pending, user: user2)
   end
 
   let!(:granted_for_other_user_and_method) do
-    create(:authorization, :granted)
+    create(:authorization, :granted, user: user3)
+  end
+
+  let!(:external_organization_authorization) do
+    create(:authorization)
   end
 
   shared_examples_for "a correct usage of the query" do
     subject { described_class.new(parameters).query.to_a }
 
     it { is_expected.to match_array(expectation) }
+    it { is_expected.not_to include(external_organization_authorization) }
   end
 
   describe "no filtering" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: nil, name: nil, granted: nil }
+        { organization: organization, user: nil, name: nil, granted: nil }
       end
 
       let(:expectation) do
@@ -63,7 +71,7 @@ describe Decidim::Verifications::Authorizations do
   describe "granted only" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: nil, name: nil, granted: true }
+        { organization: organization, user: nil, name: nil, granted: true }
       end
 
       let(:expectation) do
@@ -80,7 +88,7 @@ describe Decidim::Verifications::Authorizations do
   describe "pending only" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: nil, name: nil, granted: false }
+        { organization: organization, user: nil, name: nil, granted: false }
       end
 
       let(:expectation) do
@@ -96,7 +104,7 @@ describe Decidim::Verifications::Authorizations do
   describe "by method" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: nil, name: name, granted: nil }
+        { organization: organization, user: nil, name: name, granted: nil }
       end
 
       let(:expectation) do
@@ -112,7 +120,7 @@ describe Decidim::Verifications::Authorizations do
   describe "granted by method" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: nil, name: name, granted: true }
+        { organization: organization, user: nil, name: name, granted: true }
       end
 
       let(:expectation) do
@@ -124,7 +132,7 @@ describe Decidim::Verifications::Authorizations do
   describe "pending by method" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: nil, name: name, granted: false }
+        { organization: organization, user: nil, name: name, granted: false }
       end
 
       let(:expectation) do
@@ -136,7 +144,7 @@ describe Decidim::Verifications::Authorizations do
   describe "by user" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: user, name: nil, granted: nil }
+        { organization: organization, user: user, name: nil, granted: nil }
       end
 
       let(:expectation) do
@@ -152,7 +160,7 @@ describe Decidim::Verifications::Authorizations do
   describe "granted by user" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: user, name: nil, granted: true }
+        { organization: organization, user: user, name: nil, granted: true }
       end
 
       let(:expectation) do
@@ -164,7 +172,7 @@ describe Decidim::Verifications::Authorizations do
   describe "pending by user" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: user, name: nil, granted: false }
+        { organization: organization, user: user, name: nil, granted: false }
       end
 
       let(:expectation) { [pending_for_other_method] }
@@ -174,7 +182,7 @@ describe Decidim::Verifications::Authorizations do
   describe "by user and method" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: user, name: name, granted: nil }
+        { organization: organization, user: user, name: name, granted: nil }
       end
 
       let(:expectation) { [granted_for_user_and_method] }
@@ -184,7 +192,7 @@ describe Decidim::Verifications::Authorizations do
   describe "granted by user and method" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: user, name: name, granted: true }
+        { organization: organization, user: user, name: name, granted: true }
       end
 
       let(:expectation) { [granted_for_user_and_method] }
@@ -194,7 +202,7 @@ describe Decidim::Verifications::Authorizations do
   describe "pending by user and method" do
     it_behaves_like "a correct usage of the query" do
       let(:parameters) do
-        { user: user, name: name, granted: false }
+        { organization: organization, user: user, name: name, granted: false }
       end
 
       let(:expectation) { [] }